Refactor: Flatten semantic assembly module and merge small modules
- Flattened `src/semantic/assembly/` into `src/semantic/assembler.rs`
- Merged `src/parser/recursion.rs` into `src/parser/common.rs`
- Merged `src/semantic/types.rs` into `src/semantic/model.rs`
- Removed duplicate tests in `src/semantic/assembler.rs` (covered by `src/semantic/assembler_tests.rs`)
- Updated imports and references across the codebase
- Updated `.jules/razor.md` with reduction details
