+++ src/semantic/classification_tests.rs
@@ -2,7 +2,7 @@ use super::conversion::classify_assembled_statement;
 use crate::ast::Expr;
 use crate::morphology::Case;
 use crate::morphology::lexicon::BinaryOp;
-use crate::semantic::assembly::{ParticipleConstituent, VerbConstituent};
+use crate::semantic::assembler::{ParticipleConstituent, VerbConstituent};
 use crate::semantic::{
     AnalyzedExprKind, AnalyzedStatement, AssembledStatement, Constituent, GlossaType, Literal,
     Scope,
