+++ src/semantic/mod.rs
@@ -33,7 +33,7 @@
 
 #[cfg(test)]
 mod assembler_tests;
-pub mod assembly;
+pub mod assembler;
 #[cfg(test)]
 mod classification_tests;
 pub(crate) mod control_flow;
@@ -45,16 +45,14 @@ pub(crate) mod expressions;
 pub(crate) mod model;
 pub(crate) mod patterns;
 mod resolver;
-mod types;
 
 pub use crate::morphology::{DisambiguationContext, analyze_article, disambiguate, resolve_best};
-pub(crate) use assembly::Assembler;
-pub use assembly::{
+pub(crate) use assembler::Assembler;
+pub use assembler::{
     AssembledStatement, AssemblyError, Constituent, Literal, ParticipleConstituent, VerbConstituent,
 };
 pub use model::*;
 pub use resolver::*;
-pub use types::*;
 
 use crate::ast::{Expr, Program, Statement};
 use crate::errors::GlossaError;
